refactor: modularize ui components
